+++ conftest.py
@@ -2,7 +2,7 @@ import platform
 import sys
 from typing import List
 
-collect_ignore = ['templates/*']
+collect_ignore = ['setup.py']
 
 
 def pytest_cmdline_preparse(args: List[str]) -> None:
@@ -20,11 +20,12 @@ def pytest_cmdline_preparse(args: List[str]) -> None:
     # add mypy option if not pypy - so mypy will be called with setup.py install test
     # add mypy only on 3.x versions
     # mypy does not find some functions on python 3.6
-    if platform.python_implementation() != "PyPy" and sys.version_info >= (3, 5) and sys.version_info != (3, 6):  # type: ignore
-        args[:] = ["--mypy"] + args
 
-    # for python 3.x use --pycodestyle, for python 2.7 use --pep8
-    if sys.version_info <= (3, 5):
-        args[:] = ["--pep8"] + args
-    else:
-        args[:] = ["--pycodestyle"] + args
+    additional_arg: List[str]
+
+    if platform.python_implementation() != "PyPy" and sys.version_info >= (3, 5) and sys.version_info != (3, 6):  # type: ignore
+        additional_arg = ["--mypy"]
+        args[:] = additional_arg + args
+
+    additional_arg = ["--pycodestyle"]
+    args[:] = additional_arg + args