+++ rst_include/__init__.py
@@ -1,8 +1,8 @@
-# this needs to come after the module imports, otherwise circular import under windows
+__all__ = ["rst_inc", "rst_str_replace"]
 from .libs.lib_main import rst_inc
 from .libs.lib_main import rst_str_replace
 
-from . import __init__conf__                            # valid relative Import
+from . import __init__conf__
 __title__ = __init__conf__.title
 __version__ = __init__conf__.version
 __name__ = __init__conf__.name