+++ rst_include/__init__conf__.py
@@ -1,6 +1,6 @@
 name = 'rst_include'
-title = 'since You can not include files into RST files on github and PyPi, You can replace those imports with this software.'
-version = '2.0.0'
+title = 'commandline tool to resolve RST File includes'
+version = '2.0.1'
 url = 'https://github.com/bitranox/rst_include'
 author = 'Robert Nowotny'
 author_email = 'bitranox@gmail.com'
@@ -12,9 +12,9 @@ def print_info() -> None:
 
 Info for rst_include:
 
-    since You can not include files into RST files on github and PyPi, You can replace those imports with this software.
+    commandline tool to resolve RST File includes
 
-    Version : 2.0.0
+    Version : 2.0.1
     Url     : https://github.com/bitranox/rst_include
     Author  : Robert Nowotny
     Email   : bitranox@gmail.com""")