+++ rst_include/rst_include.py
@@ -1,13 +1,9 @@
 # STDLIB
-import errno
 import importlib
 import importlib.util
-import pathlib
+import pathlib3x as pathlib
 import sys
-from typing import Dict, Tuple, Union, IO
-
-# EXT
-import click
+from typing import Tuple, Union, IO
 
 # OWN
 import lib_log_utils                # type: ignore
@@ -80,7 +76,7 @@ def import_module_from_file(module_fullpath: Union[pathlib.Path, str], reload: b
     module_fullpath = pathlib.Path(module_fullpath)
 
     if not module_fullpath.suffix == '.py':
-        module_fullpath = pathlib.Path(str(module_fullpath) + '.py')
+        module_fullpath = module_fullpath.with_suffix('.py')
 
     module_name = module_fullpath.stem
 
@@ -157,82 +153,6 @@ def adjust_cli_parameters(target: str, quiet: bool, inplace: bool) -> Tuple[str,
     return target, quiet
 
 
-def main() -> None:
-    """
-    >>> main()
-    """
-    pass
-
-
-def info() -> None:
-    """
-    >>> info()  # doctest: +ELLIPSIS +NORMALIZE_WHITESPACE
-    Info for ...
-
-    """
-    __init__conf__.print_info()
-
-
-@click.group(help=__init__conf__.title, context_settings=CLICK_CONTEXT_SETTINGS)
-@click.version_option(version=__init__conf__.version,
-                      prog_name=__init__conf__.shell_command,
-                      message='{} version %(version)s'.format(__init__conf__.shell_command))
-def cli_main() -> None:         # pragma: no cover
-    main()                              # pragma: no cover
-
-
-@cli_main.command('info', context_settings=CLICK_CONTEXT_SETTINGS)
-def cli_info() -> None:                     # pragma: no cover
-    """ get program informations """
-    info()                                  # pragma: no cover
-
-
-@cli_main.command('include', context_settings=CLICK_CONTEXT_SETTINGS)
-@click.argument('source', type=click.Path(exists=True, file_okay=True, dir_okay=False, readable=True, resolve_path=True, allow_dash=True))
-@click.argument('target', type=click.Path(exists=False, file_okay=True, dir_okay=False, readable=False, resolve_path=True, allow_dash=True))
-@click.option('-q', '--quiet', is_flag=True)
-@click.option('-i', '--inplace', is_flag=True, help='TARGET is discarded, use "-"')
-@click.option('-e', '--source_encoding', default='utf8-sig', help='default: utf8-sig')
-@click.option('-E', '--target_encoding', default='utf8', help='default: utf8')
-def cli_include(source: str, target: str, quiet: bool, inplace: bool, source_encoding: str, target_encoding: str) -> None:  # pragma: no cover
-    """ \b
-        include the include files,
-        use "-" for stdin as SOURCE and
-            "-" for stdout as TARGET """
-    include(source=source, target=target, quiet=quiet, inplace=inplace, source_encoding=source_encoding, target_encoding=target_encoding)   # pragma: no cover
-
-
-@cli_main.command('replace', context_settings=CLICK_CONTEXT_SETTINGS)
-@click.argument('source', type=click.Path(exists=True, file_okay=True, dir_okay=False, readable=True, resolve_path=True, allow_dash=True))
-@click.argument('target', type=click.Path(exists=False, file_okay=True, dir_okay=False, readable=False, resolve_path=True, allow_dash=True))
-@click.argument('str_pattern')
-@click.argument('str_replace')
-@click.option('-c', '--count', type=int, default=-1)
-@click.option('-q', '--quiet', is_flag=True)
-@click.option('-i', '--inplace', is_flag=True, help='TARGET is discarded, use "-"')
-@click.option('-e', '--source_encoding', default='utf8-sig', help='default: utf8-sig')
-@click.option('-E', '--target_encoding', default='utf8', help='default: utf8')
-def cli_replace(source: str, target: str, str_pattern: str, str_replace: str, count: int,
-                quiet: bool, inplace: bool, source_encoding: str, target_encoding: str) -> None:  # pragma: no cover
-    """ replace <str_pattern> with <str_replace> <count> times """
-    replace(source=source, target=target, str_pattern=str_pattern, str_replace=str_replace, count=count,
-            quiet=quiet, inplace=inplace, source_encoding=source_encoding, target_encoding=target_encoding)
-
-
 # entry point if main
 if __name__ == '__main__':
-    try:
-        cli_main()
-    except FileNotFoundError:
-        # see https://www.thegeekstuff.com/2010/10/linux-error-codes for error codes
-        # No such file or directory
-        sys.exit(errno.ENOENT)      # pragma: no cover
-    except FileExistsError:
-        # File exists
-        sys.exit(errno.EEXIST)      # pragma: no cover
-    except TypeError:
-        # Invalid Argument
-        sys.exit(errno.EINVAL)      # pragma: no cover
-        # Invalid Argument
-    except ValueError:
-        sys.exit(errno.EINVAL)      # pragma: no cover
+    print('this is a library only, the executable is named rst_include_cli.py')