+++ tests/local_testscripts/lib_bash_functions.sh
@@ -1,23 +1,15 @@
 #!/bin/bash
 save_dir="$PWD"
-own_dir="$( cd "$(dirname "${BASH_SOURCE[0]}")" || exit && pwd -P )" # this gives the full path, even for sourced scripts
-
+own_dir="$(cd "$(dirname "${BASH_SOURCE[0]}")" || exit && pwd -P)" # this gives the full path, even for sourced scripts
 
 sleeptime_on_error=5
 sudo_askpass="$(command -v ssh-askpass)"
 export SUDO_ASKPASS="${sudo_askpass}"
-export NO_AT_BRIDGE=1                                     # get rid of (ssh-askpass:25930): dbind-WARNING **: 18:46:12.019: Couldn't register with accessibility bus: Did not receive a reply.
+export NO_AT_BRIDGE=1                        # get rid of (ssh-askpass:25930): dbind-WARNING **: 18:46:12.019: Couldn't register with accessibility bus: Did not receive a reply.
 
-tests_dir="${own_dir}"
-project_root_dir="$(dirname "${tests_dir}")"              # one level up
-# if we have other Projects stored in that directory, we can import them without installing, otherwise not harmful
-above_project_root_dir="$(dirname "${project_root_dir}")" # one level up
-export PYTHONPATH="${above_project_root_dir}":"${PYTHONPATH}"
-
-# if we have other Projects stored in that directory, we can import them without installing, otherwise not harmful
-# this we might need for rotek intern development - but then commandline registration will fail - keep this as a reminder :
-# two_above_project_root_dir="$(dirname "${above_project_root_dir}")" # one level up
-# export PYTHONPATH="${two_above_project_root_dir}":"${PYTHONPATH}"
+tests_dir="$(dirname "${own_dir}")"          # one level up
+project_root_dir="$(dirname "${tests_dir}")" # one level up
+export PYTHONPATH="${project_root_dir}:/media/srv-main-softdev/rotek-apps/lib:${PYTHONPATH}"
 
 function install_or_update_lib_bash() {
   if [[ ! -f /usr/local/lib_bash/install_or_update.sh ]]; then
@@ -56,26 +48,24 @@ function clean_caches() {
 }
 
 function install_virtualenv_debian() {
-  clr_green "installing virtualenv"
-  sudo apt-get install python3-virtualenv
+  if ! is_package_installed python3-virtualenv; then
+    banner "python3-virtualenv is not installed, I will install it for You"
+    wait_for_enter
+    install_package_if_not_present python3-virtualenv
+  fi
 }
 
-function install_requirements() {
+function install_test_requirements() {
   # this should be already installed, but it happens that pycharm ide venv does not have it
-  clr_green "install_requirements"
-  sudo chmod -R 0777 ~/.eggs    # make already installed eggs accessible, just in case they were installed as root
+  clr_green "installing/updating pip, setuptools, wheel"
+  sudo chmod -R 0777 ~/.eggs # make already installed eggs accessible, just in case they were installed as root
 
   python3 -m pip install --upgrade pip
   python3 -m pip install --upgrade setuptools
   python3 -m pip install --upgrade wheel
 
-  if test -f "${project_root_dir}/requirements.txt"; then
-    python3 -m pip install --upgrade -r "${project_root_dir}/requirements.txt"
-  else
-    clr_red "requirements.txt not found"
-  fi
-
   if test -f "${project_root_dir}/requirements_test.txt"; then
+    clr_green "installing/updating test requirements from \"requirements_test.txt\""
     python3 -m pip install --upgrade -r "${project_root_dir}/requirements_test.txt"
   else
     clr_red "requirements_test.txt not found"
@@ -83,9 +73,9 @@ function install_requirements() {
 }
 
 function install_dependencies() {
-  clr_green "installing dependencies"
+  banner "installing dependencies"
   install_virtualenv_debian
-  install_requirements
+  install_test_requirements
 }
 
 function delete_virtual_environment() {
@@ -99,15 +89,18 @@ function install_clean_virtual_environment() {
 }
 
 function cleanup() {
+  trap '' 2 # disable Ctrl+C
   delete_virtual_environment
   clean_caches
   cd "${save_dir}" || exit
+  trap 2 # enable Ctrl+C
 }
 
-function pytest_codestyle_mypy() {
-  my_banner "pytest --pycodestyle --mypy"
-  if ! python3 -m pytest "${project_root_dir}" --disable-warnings; then
-    my_banner_warning "pytest --pycodestyle --mypy ERROR"
+function run_pytest() {
+  # run pytest, accepts additional pytest parameters like --disable-warnings and so on
+  my_banner "running pytest with settings from pytest.ini, mypy.ini and conftest.py"
+  if ! python3 -m pytest "${project_root_dir}" /media/srv-main-softdev/rotek-apps/lib/bitranox/pathlib3x/pathlib3x "$@"; then
+    my_banner_warning "pytest ERROR"
     beep
     sleep "${sleeptime_on_error}"
     return 1
@@ -116,7 +109,8 @@ function pytest_codestyle_mypy() {
 
 function mypy_strict() {
   my_banner "mypy strict"
-  if ! python3 -m mypy "${project_root_dir}" --strict --warn-unused-ignores --implicit-reexport --follow-imports=silent; then
+  if ! python3 -m mypy "${project_root_dir}" /media/srv-main-softdev/rotek-apps/lib/bitranox/pathlib3x/pathlib3x \
+                       --strict --warn-unused-ignores --implicit-reexport --follow-imports=silent; then
     my_banner_warning "mypy strict ERROR"
     beep
     sleep "${sleeptime_on_error}"
@@ -126,7 +120,8 @@ function mypy_strict() {
 
 function mypy_strict_with_imports() {
   my_banner "mypy strict including imports"
-  if ! python3 -m mypy "${project_root_dir}" --strict --warn-unused-ignores --implicit-reexport --follow-imports&>/dev/null; then
+  if ! python3 -m mypy "${project_root_dir}" /media/srv-main-softdev/rotek-apps/lib/bitranox/pathlib3x/pathlib3x \
+                       --strict --warn-unused-ignores --implicit-reexport --follow-imports=normal; then
     my_banner_warning "mypy strict including imports ERROR"
     beep
     sleep "${sleeptime_on_error}"
@@ -165,9 +160,8 @@ function test_commandline_interface_venv() {
   # this will fail if rotek lib directory is in the path - keep this as a reminder
   my_banner "test commandline interface on virtual environment"
 
-  registered_shell_command=$(python3 "${project_root_dir}/project_conf.py" get_shell_command)
-  clr_green "issuing command : $HOME/venv/bin/${registered_shell_command} -v"
-  if ! "$HOME/venv/bin/${registered_shell_command}" --version; then
+  clr_green "issuing command : $HOME/venv/bin/rst_include --version"
+  if ! "$HOME/venv/bin/rst_include" --version; then
     my_banner_warning "test commandline interface on virtual environment ERROR"
     beep
     sleep "${sleeptime_on_error}"